requirements.txt
django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
